🚑 fix: separate invitation client and server code
